requirements.txt
crewai
langchain
langchain_openai
langchain-community
langchain_chroma
langchain-core
faiss-cpu
sse_starlette
python-dotenv
requests
pydantic
psycopg2-binary
tenacity
crewai-tools
onnxruntime
transformers
torch
librosa
newspaper3k
python-magic
PyYAML
feedparser
beautifulsoup4
fastapi
uvicorn
langserve[all]
pymongo
apify
apscheduler

